split sign form in 2 components
